Refactor Oculus Layers and SwapChain code to different files
